+++ src/main/java/co/infoclinic/term/snomedct/repository/SchemeRepository.java
@@ -1,5 +1,7 @@
 package co.infoclinic.term.snomedct.repository;
 
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.transaction.annotation.Transactional;
@@ -13,23 +15,30 @@ import co.infoclinic.term.snomedct.model.entity.Scheme;
 public interface SchemeRepository extends JpaRepository<Scheme, String> {
 
 	/**
-	 * 가장 최신버전의 스키마 조회
-	 * @return
+	 * 모든 릴리즈(International + Extension) version 내림차순 조회
 	 */
 	@Query(value = "SELECT * " +
 				   "FROM SCHEME " +
+				   "ORDER BY TO_DATE(SUBSTRING(VERSION, 2), 'YYYYMMDD') DESC", nativeQuery = true)
+	List<Scheme> findAllOrderByVersionDesc();
+
+
+	/**
+	 * International Edition 최신 Scheme (EXTENSION_NAME IS NULL)
+	 */
+	@Query(value = "SELECT * " +
+				   "FROM SCHEME " +
+				   "WHERE EXTENSION_NAME IS NULL " +
 				   "ORDER BY TO_DATE(SUBSTRING(VERSION, 2), 'YYYYMMDD') DESC " +
 				   "LIMIT 1", nativeQuery = true)
 	Scheme findLatest();
-	
-	
+
+
 	/**
-	 * 버전의 스키마 조회
-	 * @param version
-	 * @return
+	 * version으로 Scheme 조회
 	 */
 	@Query(value = "SELECT * " +
-			   	   "FROM SCHEME " +
-			   	   "WHERE VERSION = ?1", nativeQuery = true)
+				   "FROM SCHEME " +
+				   "WHERE VERSION = ?1", nativeQuery = true)
 	Scheme findByVersion(String version);
 }